Remove some unnecessary using directives
* Remove unnecessary casts from [PmdFlags.cs, PmdTargetType.cs]
* Address CS8618 via Array.Empty [PmdTargetType.cs]
* Remove unnecessary nullable from BinaryWriter call [PmdTargetType.cs]

+++ Libellus Library/Event/Types/Frame/PmdFlags.cs
@@ -29,8 +29,8 @@ namespace LibellusLibrary.Event.Types.Frame
 
         internal void WriteData(BinaryWriter writer)
         {
-            writer?.Write((ushort)FlagNumber);
-            writer?.Write((ushort)CmpValue);
+            writer?.Write(FlagNumber);
+            writer?.Write(CmpValue);
             writer?.Write((ushort)FlagType);
             writer?.Write((ushort)GFlagType);
         }
